+++ gen/tests/test_validation.py
@@ -1,5 +1,7 @@
 import json
 
+import pkg_resources
+
 import gen
 from gen.tests.utils import make_arguments, true_false_msg, validate_error, validate_success
 
@@ -139,6 +141,34 @@ def test_invalid_mesos_dns_set_truncate_bit():
         {'mesos_dns_set_truncate_bit': 'foo'},
         'mesos_dns_set_truncate_bit',
         true_false_msg)
+
+
+def test_validate_mesos_recovery_timeout():
+    validate_success(
+        {'mesos_recovery_timeout': '24hrs'})
+
+    validate_success(
+        {'mesos_recovery_timeout': '24.5hrs'})
+
+    validate_error(
+        {'mesos_recovery_timeout': '2.4.5hrs'},
+        'mesos_recovery_timeout',
+        "Invalid decimal format.")
+
+    validate_error(
+        {'mesos_recovery_timeout': 'asdf'},
+        'mesos_recovery_timeout',
+        "Error parsing 'mesos_recovery_timeout' value: asdf.")
+
+    validate_error(
+        {'mesos_recovery_timeout': '9999999999999999999999999999999999999999999ns'},
+        'mesos_recovery_timeout',
+        "Value 9999999999999999999999999999999999999999999 not in supported range.")
+
+    validate_error(
+        {'mesos_recovery_timeout': '1hour'},
+        'mesos_recovery_timeout',
+        "Unit 'hour' not in ['ns', 'us', 'ms', 'secs', 'mins', 'hrs', 'days', 'weeks'].")
 
 
 def test_cluster_docker_credentials():
@@ -705,3 +735,45 @@ def test_validate_custom_checks():
             'node check names: node-check-1, node-check-2.'
         ),
     )
+
+
+def test_validate_mesos_work_dir():
+    validate_success({
+        'mesos_master_work_dir': '/var/foo',
+        'mesos_agent_work_dir': '/var/foo',
+    })
+
+    # Relative path.
+    validate_error(
+        {'mesos_master_work_dir': 'foo'},
+        'mesos_master_work_dir',
+        'Must be an absolute filesystem path starting with /',
+    )
+    validate_error(
+        {'mesos_agent_work_dir': 'foo'},
+        'mesos_agent_work_dir',
+        'Must be an absolute filesystem path starting with /',
+    )
+
+    # Empty work dir.
+    validate_error(
+        {'mesos_master_work_dir': ''},
+        'mesos_master_work_dir',
+        'Must be an absolute filesystem path starting with /',
+    )
+    validate_error(
+        {'mesos_agent_work_dir': ''},
+        'mesos_agent_work_dir',
+        'Must be an absolute filesystem path starting with /',
+    )
+
+
+def test_fault_domain_disabled():
+    arguments = make_arguments(new_arguments={
+        'fault_domain_detect_filename': pkg_resources.resource_filename('gen', 'fault-domain-detect/aws.sh')
+    })
+
+    generated = gen.generate(arguments=arguments)
+
+    assert generated.arguments['fault_domain_enabled'] == 'false'
+    assert 'fault_domain_detect_contents' not in generated.arguments